fix: API 명세에 맞추어 rating 필드의 타입을 double로 변경

+++ src/main/java/com/codeit/mission/deokhugam/book/dto/BookDto.java
@@ -15,7 +15,7 @@ public record BookDto(
     String isbn,
     String thumbnailUrl,
     int reviewCount,
-    BigDecimal rating,
+    double rating,
     LocalDateTime createdAt,
     LocalDateTime updatedAt
 ) {
